backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
tensorflow==2.13.0
opencv-python==4.8.0.74
numpy==1.24.3
Pillow==10.0.0
geopy==2.4.0
folium==0.14.0
pandas==2.0.3
scikit-learn==1.3.0
redis==5.0.0
celery==5.3.1
twilio==8.5.0
sendgrid==6.9.7
python-dateutil==2.8.2